db: infrastructure to initialize indexes.
We assume (as many tables have) an `id` field for the created_index.
This isn't quite right for sqlite3 (for postgresql it is), because it
should have `AUTOINCREMENT` after `PRIMARY KEY (id)`.  Otherwise if
we delete the final entry and re-add, we'll get a duplicate.

The <tablename>_updated_index is set as a separate var field.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>