wallet: pay attention when you update the status
This is a bit of a mess, and @rustyrussell spent some time with me
debugging my cloud database before the amazing bitcoin++ conference.

The main point is that when we want to update the status of an output
in our database, we should be careful about the state transition we're
performing to ensure we leave the database in a consistent state.

For example, updating a status from SPENT to AVAILABLE requires setting
the `spend_height` to NULL.

The big question is: how does this happen? Is there a workflow that
causes our database to enter an incorrect state? I don't know, but
what I do know is that there may be some corner cases that can put
our database into an incorrect state, and we should update our database
without making any mess that it is difficult to debug.

Co-Developed-by: Rusty Russell <rusty@rustcorp.com.au>
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>