pay: change bolt11 to invstr param
We should have do it a lot of before now. So this commit is
removing the bolt11 parameter, and adding the invstr that it is
more generic.

In fact, the bolt11 params not accept also bolt12 invoices, so
this can confuse the user. In addition, this is done for the
preparation of another work that it is adding the ability to
pay an offer or invoice request directly with the pay command.

Changelog-Changed: pay: change bolt11 to invstr param
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>